Create replication queue on startup Starting the BackgroundReplication thread
git-svn-id: https://rhino-tools.svn.sourceforge.net/svnroot/rhino-tools/experiments/dht/dht@2187 079b0acf-d9fa-0310-9935-e5ade295c882

+++ Rhino.DistributedHashTable/Remote/DistributedHashTableNodeReplication.cs
@@ -96,7 +96,7 @@ namespace Rhino.DistributedHashTable.Remote
 			{
 				foreach (var segment in segments)
 				{
-					if (actions.HasTag((int)segment))
+					if (actions.HasTag(segment))
 						continue;
 					if (MarkSegmentAsAssignedToEndpoint(actions, replicationEndpoint, segment) == false)
 						continue;
